Gegner geändert Neue DB Version 68 - GS-Werte sind nun float (bzw. double) - Verbreitung, Auftreten, Jagdreaktion und Beute hinzugefügt
git-svn-id: https://moonvega.pmhost.de/svn/meistergeister/trunk@1229 c5e1f230-26d4-41f2-a5c8-149495a0d85e

+++ MeisterGeister_Tests/Held_Tests.cs
@@ -155,7 +155,7 @@ namespace MeisterGeister_Tests
             testHeld1.Modifikatoren.Clear();
             Assert.Greater(testHeld1.LebensenergieAktuell, 10, "Mehr als 10 LeP");
             Assert.AreEqual(0, testHeld1.Modifikatoren.Count, "Keine Modifikatoren");
-            int gs = testHeld1.Geschwindigkeit;
+            double gs = testHeld1.Geschwindigkeit;
             Assert.Greater(gs, 1);
             testHeld1.LebensenergieAktuell = (int)Math.Round(testHeld1.LebensenergieMax * 1.0 / 2.0) - 1;
             Assert.AreEqual(1, testHeld1.Modifikatoren.Where(m => m is Mod.NiedrigeLebensenergieModifikator).Count() , "Ein Modifikator");
